+++ src/core/tile_mipmapper.rs
@@ -181,6 +181,9 @@ impl TileCollector for TileMipmapper<'_> {
 		if self.config.mob_spawn {
 			fs::create_dir_all(&self.config.tile_dir(TileKind::Mobspawn, level))?;
 		}
+		if self.config.contour_layer {
+			fs::create_dir_all(&self.config.tile_dir(TileKind::Contourmap, level))?;
+		}
 		if self.config.block_textures.is_some() {
 			fs::create_dir_all(&self.config.tile_dir(TileKind::Textured, level))?;
 		}
@@ -251,6 +254,14 @@ impl TileCollector for TileMipmapper<'_> {
 				processed: 0,
 			}
 		};
+		let contour_stat = if self.config.contour_layer {
+			self.render_mipmap::<image::Rgba<u8>>(TileKind::Contourmap, level, coords, prev)?
+		} else {
+			MipmapStat {
+				total: 0,
+				processed: 0,
+			}
+		};
 		let textured_stat = if self.config.block_textures.is_some() {
 			self.render_mipmap::<image::Rgba<u8>>(TileKind::Textured, level, coords, prev)?
 		} else {
@@ -265,6 +276,7 @@ impl TileCollector for TileMipmapper<'_> {
 			+ biome_stat
 			+ cave_stat
 			+ mob_spawn_stat
+			+ contour_stat
 			+ textured_stat)
 	}
 }